Replace Makefile PY/MANAGE variables with a uv entry point
The variables saved little and hid what each recipe actually runs. Instead,
declare a `manage` console script in pyproject.toml so management commands
are spelled out in full as `uv run manage <command>`, both in the Makefile
and when typed by hand.

The entry point needs an importable module, so the body of manage.py moves
to djangovue/cli.py and manage.py becomes a shim that calls it. Running
`python manage.py <command>` behaves exactly as before, which keeps the
Dockerfile and scripts/e2e_server_smoke.sh working unchanged.

typecheck also drops to `uv run mypy`, and status spells out `uv run python`.

Verified: make verify passes from a clean tree (mypy now covers 14 files
including the new module), `uv run manage check` and `uv run python
manage.py check` behave identically, the console script installs into
.venv/bin/manage, and `uv sync --frozen --no-dev` still succeeds so the
Docker build is unaffected (uv.lock needed no update).
