+++ manage.py
@@ -1,25 +1,11 @@
 #!/usr/bin/env python
-"""Django command-line utility for administrative tasks."""
+"""Django command-line utility for administrative tasks.
 
-import os
-import sys
+The implementation lives in djangovue/cli.py so that it can also be installed
+as the ``manage`` console script; see ``[project.scripts]`` in pyproject.toml.
+"""
 
-
-def main() -> None:
-    """Run Django management commands from the command line."""
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "djangovue.settings")
-    try:
-        from django.core.management import execute_from_command_line
-    except ImportError as exc:
-        # The above import may fail for some other reason. Ensure that the
-        # issue is really that Django is missing.
-        raise ImportError(
-            "Couldn't import Django. Are you sure it's installed and "
-            "available on your PYTHONPATH environment variable? Did you "
-            "forget to activate a virtual environment?"
-        ) from exc
-    execute_from_command_line(sys.argv)
-
+from djangovue.cli import main
 
 if __name__ == "__main__":
     main()
